requirements.txt
# Core Framework Dependencies
streamlit>=1.46.1
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
starlette>=0.47.1

# Data Processing
pandas>=2.3.1
numpy>=2.3.1
python-dateutil>=2.9.0

# Database and Storage
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.13.0

# AWS SDK and Services
boto3>=1.34.0
botocore>=1.34.0
aioboto3>=12.3.0
opensearch-py>=2.4.0

# HTTP Clients
httpx>=0.25.0
requests>=2.32.4
aiofiles>=23.2.1

# Natural Language Processing
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
sentence-transformers>=2.2.0
transformers>=4.35.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Logging and Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Async Support
asyncio>=3.4.3
asyncpg>=0.29.0

# Data Validation and Serialization
jsonschema>=4.24.0
marshmallow>=3.20.0

# Streamlit Extensions
streamlit-chat>=0.1.1
streamlit-authenticator>=0.2.3

# Utility Libraries
click>=8.2.1
tenacity>=9.1.2
python-multipart>=0.0.6

# Production Dependencies
gunicorn>=21.2.0
supervisor>=4.2.5

# Monitoring and Observability
prometheus-client>=0.19.0
